libs/python/clients/resources/py-minio/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "pyminio" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 --cov-report html:'../../../../../coverage/libs/python/clients/resources/py-minio/html' --cov-report xml:'../../../../../coverage/libs/python/clients/resources/py-minio/coverage.xml' --html='../../../../../reports/libs/python/clients/resources/py-minio/unittests/html/index.html' --junitxml='../../../../../reports/libs/python/clients/resources/py-minio/unittests/junit.xml'"

[tool.poetry]
name = "python-clients-resources-py-minio"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "pyminio"

  [tool.poetry.dependencies]
  python = "^3.10"
  minio = "^7.2.7"

    [tool.poetry.dependencies.shared-py-log]
    path = "../../../shared/py-log"
    develop = true

    [tool.poetry.dependencies.python-service-discovery]
    path = "../../../service-discovery"
    develop = true


[tool.poetry.group.dev.dependencies]
respx = "^0.21.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
